requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
inquirerpy==0.3.4
joblib==1.4.2
numpy==1.26.4
pfzy==0.3.4
Pillow==9.5.0
prompt_toolkit==3.0.45
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
setuptools==69.5.1
threadpoolctl==3.5.0
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.43.0
